requirements.txt
pydantic>=2.0
httpx>=0.25
beautifulsoup4>=4.12
lxml>=4.9
supabase>=2.0
python-dotenv>=1.0
